Correct target branch and Python workflow in CI (#4)
* Initialize directory/package structure

This commit initializes the package structure. This should allow `pip
install -e .` to work successfully, as well as enforcing the initial
linting (`black` and `flake8`). It does _not_ add tests yet.

* Correct Github Actions target branch

The workflow needs to point to the `main` branch, not the `master`
branch.

* Correct runner Python version

3.10 and 3.1 cannot be disambiguated when read as strings, so Python
version must now be passed as a string

* Remove flake test

We'll add this back when things are more substantial; for now blocking
unused imports is very inconvenient.