Add initial testing scaffolding (#5)
* Initialize directory/package structure

This commit initializes the package structure. This should allow `pip
install -e .` to work successfully, as well as enforcing the initial
linting (`black` and `flake8`). It does _not_ add tests yet.

* Add first test

The first test is simply enforcing return types. This may not be the
test we actually want, but it proivdes an initial structure.