Tweak algebraic connectivity to pass tests (#23)
* Initialize directory/package structure

This commit initializes the package structure. This should allow `pip
install -e .` to work successfully, as well as enforcing the initial
linting (`black` and `flake8`). It does _not_ add tests yet.

* Implement Karrer rewiring algorithm

* autoformat committed code

* Tweak algebraic connectivity to pass tests

This makes three changes that shouldn't change the behavior of
`AlgebraicConnectivity` much but make things easier for the tests.

1. It swaps the order of the `nx.is_connected()` and `nx.is_directed()`
tests, because `nx.is_connected()` is not implemented for DiGraphs.
2. It sets the parameter `directed` to True by default, so that we don't
have to pass that argument in the tests.
3. Instead of using `to_undirected()`, it uses the `nx.Graph()`
constructor to make the graph undirected. For reasons that aren't 100%
clear to me, the former approach was creating a frozen graph, which
errored when you tried to remove edges from it.

* add directed=False conditional